Optimize `Endpoint#call` for the common use case
In most cases we need to send a SOAP request and access the resulting response
data. We rarely care about the full SOAP data (ie `Envelope`, `Body`) or the
HTTP headers.

With this change `Endpoint#call` returns a Hash of the data we usually care
about when talking to E-conomic. For the special cases where more fidelity is
needed, `call` yields the entire Response object to its block.

A side effect of this is we no longer leak Savon::SOAP::Response objects
throughout R-conomic.