Remove unnecessary to_i check
It was basically there to ensure we caught all instances of

    :handle => {:id => {:id => 42}}

as we changed the code around to use Handles. It is no longer necessary.

It is also slighty wrong, as we now support keys with String values. While they
do respond to to_i we don't actually call to_i on them.