Make it possible to pass both headers and cookies
So it seems there's a bug in Savon 2.11.1 where the `Cookie` header is not used
in requests if you also specify a `globals[:headers]`. Since we need to push the
`X-EconomicAppIdentifier` in headers AND also need cookies to ensure we stay
logged in, we're stuffed.

This surfaces as a

    `raise_soap_and_http_errors!': (soap:Client)
    Economic.Api.Exceptions.AuthenticationException(E02250): Not logged in -
    could not resolve authenticationContext. This may be caused by invalid
    session cookie header. Read more at: https://www.e-conomic.com/developer
    /about-the-soap-api. (id=74eeddad-ad51-40fb-8f9b-bbb6be81f029)
    (Savon::SOAPFault)

when using `connect_with_credentials`. As far as I can tell,
`connect_with_token` is not affected.

The bug in Savon has been fixed upstream, but there has been no release of Savon
for ages that we can actually rely on. So, we're even more stuffed.

Monkeypatching Savon with a working instance of `Savon::SOAPRequest#build` that
doesn't clobber the `Cookie` header seems like the lesser evil here.

The alternative, I guess, would be to simply deprecate the entire
`connect_with_credentials` flow, which pretty much falls in line with E-conomics
official stance.

I am not quite ready to go down that route just yet, especially not without
having a working state that r-conomic users can rely on if they need to continue
using `connect_with_credentials`.